de/infra/airflow/requirements.txt
polars-lts-cpu>=1.32.0
pyarrow>=21.0.0
pytz>=2025.2
requests>=2.32.4
python-dotenv>=1.1.1
clickhouse-connect>=0.8.18
psycopg2-binary>=2.9.10
redis>=6.2.0
httpx>=0.28.1
pydantic>=2.11.7
pydantic-settings>=2.10.1
structlog>=25.4.0
python-ulid>=3.1.0